% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar-chart.R
\name{group_severity}
\alias{group_severity}
\title{Create sunburst for MSNI components}
\usage{
group_severity(df, group = "group", group_order = NULL,
  group_labels = NULL, index = "msni", index_max = 5,
  weights = NULL, plot_name = "group_severity.pdf", path = "")
}
\arguments{
\item{df}{MSNA data}

\item{group}{column name to group data by, e.g. population type}

\item{group_order}{character vector of values to order groups by in graph, if NULL, alphabetical}

\item{group_labels}{character vector of labels for graph}

\item{index}{column name for scoring index to graph with}

\item{index_max}{max value index can take, either 4 or 5}

\item{weights}{column name for weights column in dataset}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}
}
\description{
Create sunburst for MSNI components
}
