% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection.R
\name{index_intersections}
\alias{index_intersections}
\title{Create intersection graph for indices in certain range of values}
\usage{
index_intersections(df, lsg = c("education_lsg", "shelter_lsg",
  "fsl_lsg", "health_lsg", "nutrition_lsg", "protection_lsg", "wash_lsg"),
  lsg_labels = c("Education", "Shelter", "Food", "Health", "Nutrition",
  "Protection", "WASH"),
  y_label = "\% in need per combination of sectors",
  index_filter = c(3, 4, 5), weighting_function = NULL,
  nintersects = 12, exclude_unique = T, print_plot = F,
  plot_name = "intersection", path = NULL)
}
\arguments{
\item{df}{data frame}

\item{lsg}{character vector of all column names of LSG indices}

\item{lsg_labels}{character vector of all labels for LSG indices}

\item{index_filter}{numeric vector where index will get 1s if in the filter, and 0s otherwise}

\item{weighting_function}{function for weighting the data frame}

\item{nintersects}{number of intersections to include in the plot}

\item{exclude_unique}{whether the set intersections should include singular sete. Note that if this is set to True, the total set size on the left will be wrong}

\item{print_plot}{logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}
}
\description{
Create intersection graph for indices in certain range of values
}
