% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{non_zero_round}
\alias{non_zero_round}
\title{Helper function for sunbursts to not round to 0 for plot_ly input, since it only takes integer values}
\usage{
non_zero_round(x)
}
\description{
Helper function for sunbursts to not round to 0 for plot_ly input, since it only takes integer values
}
