% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.R
\name{radar_graph}
\alias{radar_graph}
\title{Create radar graph for LSG scores}
\usage{
radar_graph(df, lsg = c("education_lsg", "shelter_lsg", "fsl_lsg",
  "health_lsg", "nutrition_lsg", "protection_lsg", "wash_lsg"),
  lsg_labels = c("Education in\\nEmergency",
  "Emergency Shelter\\nand NFI", "Food Security and\\nAgriculture",
  "Health", "Nutrition", "Protection", "WASH"), group = NULL,
  group_labels = NULL, weighting_function = NULL,
  legend_position = "right")
}
\arguments{
\item{df}{MSNA data}

\item{lsg}{character vector of all column names of LSG indices}

\item{lsg_labels}{character vector of all labels for LSG indices}

\item{group}{column name to group data by, e.g. population type}

\item{group_labels}{character vector of labels for graph}

\item{weighting_function}{weighting function used to weight MSNA dataset}
}
\description{
Create radar graph for LSG scores
}
