% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar-chart.R
\name{severity_bar_chart}
\alias{severity_bar_chart}
\title{Create stacked bar chart of severity scores for MSNI or other indices}
\usage{
severity_bar_chart(df, group = "group", group_order = NULL,
  group_labels = NULL, index = "msni", index_max = 4,
  index_type = "msni", weighting_function = NULL, print_plot = F,
  plot_name = "severity_bar_chart", path = NULL)
}
\arguments{
\item{df}{MSNA data}

\item{group}{column name to group data by, e.g. population type}

\item{group_order}{character vector of values to order groups by in graph, if NULL, alphabetical}

\item{group_labels}{character vector of labels for graph}

\item{index}{column name for scoring index to graph with}

\item{index_max}{max value index can take, either 4 or 5}

\item{index_type}{character vector specifying either "msni" or "lsg", to determine labels on scores}

\item{weighting_function}{weighting function used to weight MSNA dataset}

\item{print_plot}{Logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}
}
\description{
Create stacked bar chart of severity scores for MSNI or other indices
}
