% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/severity-lines.R
\name{severity_lines}
\alias{severity_lines}
\title{Create line graph to show households by # of sector severity scores 3 and over}
\usage{
severity_lines(df, lsg = c("education_lsg", "shelter_lsg", "fsl_lsg",
  "health_lsg", "nutrition_lsg", "protection_lsg", "wash_lsg"),
  group = NULL, group_order = NULL, group_labels = NULL,
  weighting_function = NULL, print_plot = F,
  plot_name = "severity_lines", path = NULL)
}
\arguments{
\item{df}{MSNA data}

\item{lsg}{character vector of all column names of LSG indices}

\item{group}{column name to group data by, e.g. population type}

\item{group_order}{character vector of how to order the graph, defaults to alphabetical ordering}

\item{weighting_function}{weighting function used to weight MSNA dataset}

\item{print_plot}{logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}

\item{lsg_labels}{character vector of all labels for LSG indices}
}
\description{
Create line graph to show households by # of sector severity scores 3 and over
}
