% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst}
\alias{sunburst}
\title{Creates sunburst from binary columns}
\usage{
sunburst(df, cols, labels, parents, weights = NULL, colors, na.rm = T)
}
\arguments{
\item{df}{data frame}

\item{cols}{character vector of binary columns to calculate groups from}

\item{labels}{character vector of labels for each column name}

\item{parents}{character vector of the parent column for each column. The top value should have blank value for parent, ""}

\item{colors}{character vector of hex colors for branches, should ideally be a color specified for each 2nd level branch}

\item{na.rm}{if TRUE, ignore NAs in summation of binary columns}
}
\value{
plot_ly plot object
}
\description{
Creates sunburst from binary columns
}
