% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_msni}
\alias{venn_msni}
\title{Create venn diagram of living standards gaps >= 3 and capacity gaps >= 3}
\usage{
venn_msni(df, lsg = c("education_lsg", "shelter_lsg", "fsl_lsg",
  "health_lsg", "nutrition_lsg", "protection_lsg", "wash_lsg"),
  capacity_gaps = "capacity_gaps", weighting_function = NULL,
  print_plot = F, plot_name = "venn", path = NULL)
}
\arguments{
\item{df}{MSNA data}

\item{lsg}{character vector of all column names of LSG indices}

\item{capacity_gaps}{column name for capacity/coping index}

\item{weighting_function}{function for weighting the data frame}

\item{print_plot}{logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}
}
\description{
Create venn diagram of living standards gaps >= 3 and capacity gaps >= 3
}
