% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msni.R
\name{msni}
\alias{msni}
\title{Calculate the MSNI for 2019}
\usage{
msni(education_lsg, fsl_lsg, health_lsg, protection_lsg, shelter_lsg,
  wash_lsg, capacity_gaps, impact)
}
\arguments{
\item{capacity_gaps}{vector of capacity gap scores (1-4)}

\item{impact}{vector of impact scores (1-4)}

\item{*_lsg}{vector of sectoral living standard gap scores (1-4)}
}
\value{
vector of msni scores
}
\description{
Calculate the MSNI for 2019
}
