% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radar.R
\name{radar_graph}
\alias{radar_graph}
\title{Create radar graph for LSG scores}
\usage{
radar_graph(
  df,
  lsg = c("education_lsg", "shelter_lsg", "fsl_lsg", "health_lsg", "nutrition_lsg",
    "protection_lsg", "wash_lsg"),
  lsg_labels = c("Education in\\nEmergency", "Emergency Shelter\\nand NFI",
    "Food Security and\\nAgriculture", "Health", "Nutrition", "Protection", "WASH"),
  group = NULL,
  group_order = NULL,
  group_labels = NULL,
  weighting_function = NULL,
  legend_position = "right",
  legend_text_size = 10,
  label_text_size = 4,
  print_plot = F,
  plot_name = "radar_graph",
  path = NULL,
  index_threshold = 3
)
}
\arguments{
\item{df}{MSNA data}

\item{lsg}{character vector of all column names of LSG indices}

\item{lsg_labels}{character vector of all labels for LSG indices}

\item{group}{column name to group data by, e.g. population type}

\item{group_order}{character vector of how to order the graph, defaults to alphabetical ordering}

\item{group_labels}{character vector of labels for graph, make sure they match the order for the data}

\item{weighting_function}{weighting function used to weight MSNA dataset}

\item{legend_position}{character specifying legend location, either "top", "left", "right", or "bottom"}

\item{legend_text_size}{integer specifying size of legend text}

\item{label_text_size}{integer specifying size of label text}

\item{print_plot}{logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}

\item{index_threshold}{sectoral index threshold to graph percentage greater than (default=3)}
}
\description{
Create radar graph for LSG scores
}
