% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sunburst.R
\name{sunburst_msni}
\alias{sunburst_msni}
\title{Create sunburst for MSNI components}
\usage{
sunburst_msni(df, msni = "msni", fsl_lsg = "fsl_lsg",
  health_lsg = "health_lsg", protection_lsg = "protection_lsg",
  shelter_lsg = "shelter_lsg", wash_lsg = "wash_lsg",
  capacity_gaps = "capacity_gaps", impact = "impact",
  msni_filter = c(3, 4, 5), weighting_function = NULL,
  fsl_wash_branch = F, health_prot_shelt_branch = F,
  impact_branch = F, print_plot = F, plot_name = "sunburst",
  path = NULL)
}
\arguments{
\item{df}{MSNA data}

\item{msni}{column name for MSNI}

\item{fsl_lsg}{column name for FSL index}

\item{health_lsg}{column name for health index}

\item{protection_lsg}{column name for protection index}

\item{shelter_lsg}{column name for shelter index}

\item{wash_lsg}{column name for wash index}

\item{capacity_gaps}{column name for capacity/coping index}

\item{impact}{column name for impact index}

\item{msni_filter}{filter dataset for households with only specific MSNI scores}

\item{weighting_function}{weighting function used to weight MSNA dataset}

\item{fsl_wash_branch}{if TRUE, displays 3rd branch layer of FSL/WASH scores}

\item{impact_branch}{if TRUE, displays 3rd layer of impact branch}

\item{print_plot}{logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}

\item{health_prot_shelter_branch}{if TRUE, displays 3rd branch layer for H/P/S scores}
}
\description{
Create sunburst for MSNI components
}
