% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{venn_msni}
\alias{venn_msni}
\title{Create venn diagram of living standards gaps >= 3 and capacity gaps >= 3}
\usage{
venn_msni(df, fsl_lsg = "fsl_lsg", health_lsg = "health_lsg",
  protection_lsg = NULL, shelter_lsg = "shelter_lsg",
  wash_lsg = "wash_lsg", education_lsg = "education_lsg",
  capacity_gaps = "capacity_gaps", weights = NULL, print_plot = F,
  plot_name = "venn", path = NULL)
}
\arguments{
\item{df}{MSNA data}

\item{fsl_lsg}{column name for FSL index}

\item{health_lsg}{column name for health index}

\item{protection_lsg}{column name for protection index}

\item{shelter_lsg}{column name for shelter index}

\item{wash_lsg}{column name for wash index}

\item{education_lsg}{column name for education index}

\item{capacity_gaps}{column name for capacity/coping index}

\item{weights}{column name for weights column in dataset}

\item{print_plot}{logical column indicating whether or not to save the plot to PDF}

\item{plot_name}{name to save plot with}

\item{path}{path to save plot to if not in current working directory}
}
\description{
Create venn diagram of living standards gaps >= 3 and capacity gaps >= 3
}
